server/pyproject.toml
[tool.poetry]
name = "smart-banquet-backend"
version = "0.1.0"
description = ""
authors = ["pristine-dev <infoc@pristine-minds.com>"]
license = "GNU Affero General Public License v3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "*"
django-environ = "*"
Django = ">=4.2,<4.3"
drf-writable-nested = "*"
djangorestframework = "*"
psycopg2-binary = "*"
django-cors-headers = "*"
django-storages = "*"
ipython = "*"
uwsgi = "*"
django-filter = "*"
pillow = "*"
django-reversion = "*"
django-admin-autocomplete-filter = "*"

[tool.poetry.dev-dependencies]
dacite = "*"
pytest = "*"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
log_cli = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
# NOTE: Update in .pre-commit-config.yaml as well
extend-exclude = "(__pycache__|.*snap_test_.*\\.py|.+\\/.+\\/migrations\\/.*)"

[tool.isort]
profile = "black"
multi_line_output = 3
skip = [
    "**/__pycache__",
    "**/snap_test_*.py",
    ".venv/",
    "**/migrations/*.py",
]

[tool.pyright]
exclude = [
    "**/__pycache__",
    "**/snap_test_*.py",
    ".venv/",
    "**/migrations/*.py",
]
reportMissingImports = true
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"
